refactor(dsp): remove transfer spi
